// src/api/models/job.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */

export interface Job {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  summary: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  department: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  employment_type: string | null;
  /** @nullable */
  apply_url: string | null;
  /** @nullable */
  apply_deadline: string | null;
  is_active: boolean;
  is_featured: boolean;
  /** @nullable */
  published_at: string | null;
  created_at: string;
  /** @nullable */
  created_by?: string | null;
  updated_at: string;
  /** @nullable */
  updated_by?: string | null;
}